modal refactor
